Move Settings to left panel, rework menu item layout
